requirements.txt
# Keep crewai at version 0.130.0
# Other package versions are flexible and can be changed
# Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means
crewai==0.130.0
fastapi
uvicorn
crewai
litellm
python-dotenv
pypdf